+++ panels/notification/center/notifystagingmodel.cpp
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2024 UnionTech Software Technology Co., Ltd.
+// SPDX-FileCopyrightText: 2024-2026 UnionTech Software Technology Co., Ltd.
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
@@ -197,7 +197,7 @@ QVariant NotifyStagingModel::data(const QModelIndex &index, int role) const
     } else if (role == NotifyRole::NotifyAppName) {
         return notify->appName();
     } else if (role == NotifyRole::NotifyIconName) {
-        return notify->entity().appIcon();
+        return notify->entity().appIconResolved();
     } else if (role == NotifyRole::NotifyTitle) {
         return notify->entity().summary();
     } else if (role == NotifyRole::NotifyContent) {
